TXREQ: fix bug in addr of TXREQ when there are multiple slices
