+++ src/main/scala/coupledL2/tl2tl/TL2TLCoupledL2.scala
@@ -166,18 +166,6 @@ class TL2TLCoupledL2(implicit p: Parameters) extends CoupledL2Base {
       case None =>
     }
 
-    def restoreAddress(x: UInt, idx: Int) = {
-      restoreAddressUInt(x, idx.U)
-    }
-    def restoreAddressUInt(x: UInt, idx: UInt) = {
-      if(bankBits == 0){
-        x
-      } else {
-        val high = x >> offsetBits
-        val low = x(offsetBits - 1, 0)
-        Cat(high, idx(bankBits - 1, 0), low)
-      }
-    }
     def bank_eq(set: UInt, bankId: Int, bankBits: Int): Bool = {
       if(bankBits == 0) true.B else set(bankBits - 1, 0) === bankId.U
     }
